use strict compiler flags and change to C++20

+++ src/gfx/buffer.h
@@ -7,6 +7,6 @@
 namespace gfx {
 template <typename T>
 inline const bgfx::Memory* make_data_ref(const std::vector<T>& data) {
-  return bgfx::makeRef(&data.front(), sizeof(T) * data.size());
+  return bgfx::makeRef(&data.front(), static_cast<uint32_t>(sizeof(T) * data.size()));
 }
 }  // namespace gfx
